requirements.txt
tensorflow==2.15.0
keras==2.15.0
numpy==1.24.4
pandas==1.5.3
matplotlib==3.7.1
seaborn==0.12.2
scikit-learn==1.3.0
streamlit==1.32.2
opencv-python==4.9.0.80
pillow==10.2.0